requirements.txt
PyOpenGL==3.1.0
dm-haiku==0.0.10
einops==0.6.1
flax==0.6.10
glumpy==1.2.0
healpy==1.16.5
imageio==2.31.4
imgaug==0.4.0
jax-resnet @ git+https://github.com/Ending2015a/jax-resnet.git
jaxlie @ git+https://github.com/Ending2015a/jaxlie.git
matplotlib==3.7.3
omegaconf
open3d==0.15.2
opencv-python==4.8.0.76
optax==0.1.7
pillow==10.0.1
pycocotools @ git+https://github.com/MartinSmeyer/cocoapi.git@v1.0#subdirectory=PythonAPI
pyyaml
ruamel.yaml
scikit-image==0.21.0
scikit-learn==1.3.1
tensorflow-addons==0.21.0
tensorflow-datasets==4.9.2
tensorflow-probability==0.21.0
tensorflow==2.13.1
torch==2.1.2
tqdm==4.66.1
transforms3d==0.4.1
